j'ai rajouté les deux fichiers suivants dans le .gitignore : report_cube.bbl report_cube.bgl
those are two files generated by IEEE bibliography. I don't know if it's a good idea to upload them on the server.

About the bibliography, in order to have it working you have to compile report_cub with bibtex :
bibtex report_cube
Then the two files above will be created and the bibliography will work properly.